requirements.txt
rich>=13.7
requests>=2.32
python-dotenv>=1.0
google-genai>=0.5
